requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.6
httpx==0.27.2
spacy>=3.7.0
hanlp>=2.1.0
sentence-transformers>=2.2.0
torch --index-url https://download.pytorch.org/whl/cpu
openai>=1.0.0
python-dotenv>=1.0.0
